+++ firebase.test.js
@@ -1,17 +1,20 @@
+import '@shgysk8zer0/polyfills';
 import { describe, test } from 'node:test';
 import assert from 'node:assert';
 import { signal } from './signal.test.js';
-import { getFirebaseJWK } from './firebase.js';
-
+import { MIME_TYPE } from './consts.js';
+import { getFirebasePublicKey, verifyFirebaseIdToken } from './firebase.js';
 
 describe('Firebase JWT tests', async () => {
-	const KID = '02100716fdd904e5b4d49116ff5dbdfc98999401';
-
-	// kid can change, so hard-coding this causes failing tests
-	test('Import Firebase public key', { signal, skip: true }, async () => {
-		const key = await getFirebaseJWK(KID, false, { signal, referrerPolicy: 'no-referrer', mode: 'cors' });
+	test('Import Firebase public key', { signal }, async () => {
+		const key = await getFirebasePublicKey(false, { signal, referrerPolicy: 'no-referrer', mode: 'cors', headers: { Accept: MIME_TYPE }});
 		assert.ok(key instanceof CryptoKey, '`getFirebaseJWK` should return a `CryptoKey`.');
 	});
 
-	test.todo('Cannot test Firebase Auth ID tokens without a token to test.');
+	// Set token via `export ID_TOKEN='...' to test. I'm not including an actual token in the codebase,
+	// and I cannot find a public token for testing.
+ 	test('Check Firebase ID token, if `ID_TOKEN` set in environment', { signal, skip: typeof process.env.ID_TOKEN !== 'string' }, async () => {
+		const result = await verifyFirebaseIdToken(process.env.ID_TOKEN, { signal });
+		assert.ok(! (result instanceof Error), 'Firebase id token should be valid.');
+	});
 });